fix: post rebase bugs and cleanup
